miniproject-bmsit/requirements.txt
flask
Flask-SQLAlchemy
flask-login
Flask-Session
redis
sqlalchemy-libsql